# internal/transport/http/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API
  version: 1.0.0
  description: |
    Production-oriented auth service with JWT access tokens, refresh sessions,
    email verification, password reset, OAuth login, profile management, and admin moderation endpoints.
servers:
  - url: http://localhost:8080
tags:
  - name: Health
  - name: Auth
  - name: OAuth
  - name: Users
  - name: Admin
paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      responses:
        '200':
          description: Service is healthy
          content:
            text/plain:
              schema:
                type: string
                example: ok

  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login by email or username
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh JWT using a refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token pair rotated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/auth/logout:
    post:
      tags: [Auth]
      summary: Logout by revoking a refresh session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/verify-email/request:
    post:
      tags: [Auth]
      summary: Send verification email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestEmailVerificationRequest'
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/auth/verify-email/confirm:
    post:
      tags: [Auth]
      summary: Confirm email verification by token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/password-reset/request:
    post:
      tags: [Auth]
      summary: Send password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestPasswordResetRequest'
      responses:
        '202':
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Reset password by token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /api/v1/auth/oauth/{provider}/start:
    get:
      tags: [OAuth]
      summary: Start OAuth flow
      parameters:
        - $ref: '#/components/parameters/OAuthProvider'
      responses:
        '200':
          description: OAuth authorization URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthStartResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /api/v1/auth/oauth/{provider}/callback:
    get:
      tags: [OAuth]
      summary: Handle OAuth callback
      parameters:
        - $ref: '#/components/parameters/OAuthProvider'
        - in: query
          name: code
          required: true
          schema:
            type: string
        - in: query
          name: state
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OAuth login succeeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthCallbackResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/users/me:
    get:
      tags: [Users]
      summary: Get current user profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/users/me/profile:
    put:
      tags: [Users]
      summary: Update current user profile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/admin/users/{userID}/status:
    patch:
      tags: [Admin]
      summary: Update user account status
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserStatusRequest'
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  /api/v1/admin/users/{userID}/role:
    patch:
      tags: [Admin]
      summary: Update user role
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleRequest'
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OAuthProvider:
      in: path
      name: provider
      required: true
      schema:
        type: string
        enum: [google, github, vk]
    UserID:
      in: path
      name: userID
      required: true
      schema:
        type: integer
        format: int64
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    StatusResponse:
      type: object
      properties:
        status:
          type: string
      required: [status]
    AcceptedResponse:
      type: object
      properties:
        status:
          type: string
          example: accepted
      required: [status]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          example: step.user
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: StrongPass123!
      required: [username, email, password]
    RegisterResponse:
      type: object
      properties:
        status:
          type: string
          example: created
      required: [status]
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          example: step.user
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: StrongPass123!
      required: [password]
    RefreshTokenRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    LogoutRequest:
      type: object
      properties:
        refresh_token:
          type: string
      required: [refresh_token]
    VerifyEmailRequest:
      type: object
      properties:
        token:
          type: string
      required: [token]
    RequestEmailVerificationRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    RequestPasswordResetRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required: [email]
    ResetPasswordRequest:
      type: object
      properties:
        token:
          type: string
        new_password:
          type: string
          format: password
      required: [token, new_password]
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        access_token_expires_at:
          type: string
          format: date-time
        refresh_token:
          type: string
        refresh_token_expires_at:
          type: string
          format: date-time
      required: [access_token, access_token_expires_at, refresh_token, refresh_token_expires_at]
    UpdateProfileRequest:
      type: object
      properties:
        full_name:
          type: string
        avatar_url:
          type: string
          format: uri
        status:
          type: string
        timezone:
          type: string
      required: [full_name, avatar_url, status, timezone]
    UpdateUserRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum: [user, moderator, admin]
      required: [role]
    UpdateUserStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, blocked, deactivated]
      required: [status]
    OAuthStartResponse:
      type: object
      properties:
        provider:
          type: string
          enum: [google, github, vk]
        authorization_url:
          type: string
        state:
          type: string
      required: [provider, authorization_url, state]
    OAuthCallbackResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserProfileResponse'
        tokens:
          $ref: '#/components/schemas/TokenResponse'
      required: [user, tokens]
    UserProfile:
      type: object
      properties:
        user_id:
          type: integer
          format: int64
        full_name:
          type: string
        avatar_url:
          type: string
        status:
          type: string
        timezone:
          type: string
        updated_at:
          type: string
          format: date-time
      required: [user_id, full_name, avatar_url, status, timezone, updated_at]
    UserProfileResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        role:
          type: string
          enum: [user, moderator, admin]
        account_status:
          type: string
          enum: [active, blocked, deactivated]
        profile:
          $ref: '#/components/schemas/UserProfile'
      required: [id, username, email, role, account_status, profile]